Resolution de l'erreur Internal server error: Server-only module referenced by client
