+++ src/feats/courses/functions.ts
@@ -1,10 +1,12 @@
 import { InferSelectModel, eq } from "drizzle-orm";
 import { db } from "~/common/utils/db.server";
-import { asynchronousCourses, lessons } from "../asynchronous-courses/schema";
+import { getOneAsynchronous } from "../asynchronous-courses/function";
+import { lessons } from "../asynchronous-courses/schema";
 import { classes, synchronousCourses } from "../synchronous-courses/schema";
 import { courses } from "./schema";
+import { serverOnly$ } from "vite-env-only";
 
-export type CourseCardAsync = {
+export type CourseCardAsync =  Omit<InferSelectModel<typeof lessons>, "id" | "creationTime" | "updateTime" | "title" | "description"> & {
 	chapterId: number;
 	capsuleUrl: string;
 	previewable: boolean;
@@ -30,38 +32,13 @@ export type Course = {
 
 export type CourseCard = Omit<
 	InferSelectModel<typeof classes> &
-		InferSelectModel<typeof lessons> &
-		InferSelectModel<typeof courses>,
+		InferSelectModel<typeof lessons>,
 	"id" | "creationTime" | "updateTime"
 > &
 	Course &
 	(CourseCardSync | CourseCardAsync);
 
 export async function getAllCours() {
-	const result = await db.select().from(courses);
+	const result = serverOnly$(await db.select().from(courses));
 	return result;
-}
-
-export async function getOneCours(id: number) {
-	const course = await db.select().from(courses).where(eq(courses.id, id));
-
-	const syncCourse: CourseCardSync[] = await db
-			.select()
-			.from(classes)
-			.where(eq(synchronousCourses.id, id));
-
-	let CoursInfo;
-
-	if (syncCourse.length === 0) {
-			const asyncCourseInfo: CourseCardAsync[] = await db
-					.select()
-					.from(lessons)
-					.where(eq(asynchronousCourses.id, id));
-
-			CoursInfo = [...course, ...asyncCourseInfo];
-	} else {
-			CoursInfo = [...course, ...syncCourse];
-	}
-
-	return CoursInfo;
 }
