+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-	"include": [".eslintrc.cjs", "**/*.ts", "**/*.tsx"],
+	"include": [".eslintrc.cjs", "**/*.ts", "**/*.tsx", "src/feats/courses/getOneCours.ts"],
 
 	"compilerOptions": {
 		/* || Type checking */
